Graph mis à jour règlement 2014 Suppression éléments de jeu année 2013 Ajout éléments 2014 Dessin éléments 2014 Modifications dessins table (couleur par équipe)

+++ GoBot/GoBot/Calculs/Angle.cs
@@ -133,7 +133,7 @@ namespace GoBot.Calculs
             return !(a1 == a2);
         }
 
-        public static implicit operator Angle(int angle)
+        public static implicit operator Angle(double angle)
         {
             return new Angle(angle);
         }
